Refactor harmony/responses.rs
Changes made:

  - Split the old monolithic responses.rs (~1650 lines) into 6 focused files
  - Fixed import path for RequestPipeline (routers::grpc::pipeline not harmony::pipeline)
  - Made ToolResult public for use by common/responses/streaming.rs
  - Exported ToolResult from responses/mod.rs
  - Code compiles cleanly with no clippy warnings
