+++ sgl-model-gateway/src/routers/grpc/harmony/responses/context.rs
@@ -1,0 +1,82 @@
+//! Context for Harmony Responses execution
+
+use std::sync::Arc;
+
+use tokio::sync::mpsc;
+
+use crate::{
+    data_connector::{ConversationItemStorage, ConversationStorage, ResponseStorage},
+    mcp::McpManager,
+    routers::grpc::{context::SharedComponents, pipeline::RequestPipeline},
+};
+
+/// Context for Harmony Responses execution with MCP tool support
+///
+/// Contains all dependencies needed for multi-turn Responses API execution.
+/// Cheap to clone (all Arc references).
+#[derive(Clone)]
+pub struct HarmonyResponsesContext {
+    /// Pipeline for executing Harmony requests
+    pub pipeline: Arc<RequestPipeline>,
+
+    /// Shared components (tokenizer, parsers)
+    pub components: Arc<SharedComponents>,
+
+    /// MCP manager for tool execution
+    pub mcp_manager: Arc<McpManager>,
+
+    /// Response storage for loading conversation history
+    pub response_storage: Arc<dyn ResponseStorage>,
+
+    /// Conversation storage for persisting conversations
+    pub conversation_storage: Arc<dyn ConversationStorage>,
+
+    /// Conversation item storage for persisting conversation items
+    pub conversation_item_storage: Arc<dyn ConversationItemStorage>,
+
+    /// Optional streaming sender (for future streaming support)
+    pub stream_tx: Option<mpsc::UnboundedSender<Result<String, String>>>,
+}
+
+impl HarmonyResponsesContext {
+    /// Create a new Harmony Responses context
+    pub fn new(
+        pipeline: Arc<RequestPipeline>,
+        components: Arc<SharedComponents>,
+        mcp_manager: Arc<McpManager>,
+        response_storage: Arc<dyn ResponseStorage>,
+        conversation_storage: Arc<dyn ConversationStorage>,
+        conversation_item_storage: Arc<dyn ConversationItemStorage>,
+    ) -> Self {
+        Self {
+            pipeline,
+            components,
+            mcp_manager,
+            response_storage,
+            conversation_storage,
+            conversation_item_storage,
+            stream_tx: None,
+        }
+    }
+
+    /// Create with streaming support
+    pub fn with_streaming(
+        pipeline: Arc<RequestPipeline>,
+        components: Arc<SharedComponents>,
+        mcp_manager: Arc<McpManager>,
+        response_storage: Arc<dyn ResponseStorage>,
+        conversation_storage: Arc<dyn ConversationStorage>,
+        conversation_item_storage: Arc<dyn ConversationItemStorage>,
+        stream_tx: mpsc::UnboundedSender<Result<String, String>>,
+    ) -> Self {
+        Self {
+            pipeline,
+            components,
+            mcp_manager,
+            response_storage,
+            conversation_storage,
+            conversation_item_storage,
+            stream_tx: Some(stream_tx),
+        }
+    }
+}
